Always check for duped batch before creation
Kills the "strict" variable because, as strange as it seems, "strict"
being True (and it always is from the admin command) means to *not*
validate the batch's existence.  Forces the batch to not exist before
attempting to load it, as ALL exceptions in the create code result in an
automatic purge.