iiif page viewer
These are changes to convert the OpenSeadragon viewer over to using the
IIIF tiling API in rais-image-server. I upgraded to the latest version
of OpenSeadragon (v2.1)

Old code for determining tile sizes and urls has been removed.

Work still needs to be done to convert the thumbnail and other images views
over to using iiif instead of the old tiling interface.