Add Django defaults and override in settings_base
Start settings_base with the default config for a new Django 1.9 app to
the top for easier comparison when upgrading and override with OpenONI
customizations below

When overriding lists of TEMPLATES context_processors, include original
list while commenting to disable and list additions below

Order matters for apps and middleware. Apps are currently able to be
separated similarly but may not be in the future. Middleware is mixed so
OpenONI additions are marked with comments on the right

Restore a couple middleware we had been missing (or previously removed but
not indicated)

Restore sessions (required for messages) and messages

New apps will use i18n defaults with UTC time zone etc

Remove settings which were the defaults and not overridden