Update migrations
The fields' help text values were previously set to binary "strings",
but the python 3 update changed that to a raw string, and django sees
these as different definitions.  The database schema won't actually be
changed, but future invocations of the `showmigrations` will be
confusing if we don't get this change in now.