Limit which files trigger the workflow
Add a `paths-ignore` section with a list of files for which we don't need to run the tests.

Note that there are some shell scripts arguably not currently relevant that are not excluded via this list, but that I'm allowing those to still trigger the tests, in case a later refactor should change that.